Refactor extract common drop zone functions
